Move walker propagation routines out of walker class.
In some way it's cleaner to have the propagators act on a walker as a
member function, but it's really sort of an external operation so
instead refactor these routines elsewhere.

This is a WIP to allow for different propagators to be applied to the
walkers by setting up the named routines at runtime.