Add capability to apply projector in kspace.
FFT-ing to kspace allows the kinetic energy projector to be applied
cheaply as it is diagonal in this basis. FFTs scale faster than matrix
multiplication so this should be faster than the regular direct scheme.

So far I've only seen this to be the case for larger and more dilute
systems. Should look at improving, maybe with fftw.