Don't change wavefunctions type if using reals.
It's slightly less messy to just recast a temporary to be dtype=float
rather than having complex weights throughout the rest of the code when
the wavefunction is real.

This is all because I'm still a bit confused by rfft format and how to
simply multiply this smaller matrix by the kinetic energy operator.