Initialise trial from text file.
For Chia Chen's GHF format. There is some inconsistency with the version of
the code I have regarding the ordering of the orbitals for up and down blocks.
Manually check if an orbital is spin up / down. This is important for the
format I use for the wavefunction in UHF/RHF format.